% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABU_tandem.R
\name{ABU_tandem}
\alias{ABU_tandem}
\title{Orthogonal Procrustes Problem}
\usage{
ABU_tandem(A, B, d, tol = 1e-06, max_iter = 1000)
}
\arguments{
\item{A}{An m by p matrix. The target matrix.}

\item{B}{An m by p matrix. The matrix to be aligned.}

\item{d}{A vector of length p. The diagonal matrix D.}
}
\value{
TBD
}
\description{
Orthogonal but not necessarily orthonormal Procrustes problem. Numerically solved using tandem algorithm through conjugate gradient.
}
\details{
This functions solves the orthogonal Procrustes problem.
Given matrices A and B, the orthogonal procrustes seeks a matrix U, which has orthogonal columns (U'U = D^2), such that ||A - B U||^2 is minimized.
The solution is found by a tandem algorithm using conjugate gradient method.
The current implementation assumes that the diagonal matrix D is known.
}
\examples{
ABU_tandem(A, B, d)
}
